Conflicts:
	IPython/core/interactiveshell.py
	IPython/core/prefilter.py
	IPython/extensions/parallelmagic.py
	IPython/parallel/tests/__init__.py
	IPython/utils/path.py
	IPython/utils/warn.py
	setupbase.py

+++ IPython/parallel/apps/launcher.py
@@ -21,11 +21,20 @@ import os
 import re
 import stat
 
+# signal imports, handling various platforms, versions
+
 from signal import SIGINT, SIGTERM
 try:
     from signal import SIGKILL
 except ImportError:
+    # Windows
     SIGKILL=SIGTERM
+
+try:
+    # Windows >= 2.7, 3.2
+    from signal import CTRL_C_EVENT as SIGINT
+except ImportError:
+    pass
 
 from subprocess import Popen, PIPE, STDOUT
 try:
@@ -48,15 +57,11 @@ from IPython.utils.process import find_cmd, pycmd2argv, FindCmdError
 
 from IPython.parallel.factory import LoggingFactory
 
-# load winhpcjob only on Windows
-try:
-    from .winhpcjob import (
-        IPControllerTask, IPEngineTask,
-        IPControllerJob, IPEngineSetJob
-    )
-except ImportError:
-    pass
+from .win32support import forward_read_events
 
+from .winhpcjob import IPControllerTask, IPEngineTask, IPControllerJob, IPEngineSetJob
+
+WINDOWS = os.name == 'nt'
 
 #-----------------------------------------------------------------------------
 # Paths to the kernel apps
@@ -251,9 +256,14 @@ class LocalProcessLauncher(BaseLauncher):
                 env=os.environ,
                 cwd=self.work_dir
             )
-            
-            self.loop.add_handler(self.process.stdout.fileno(), self.handle_stdout, self.loop.READ)
-            self.loop.add_handler(self.process.stderr.fileno(), self.handle_stderr, self.loop.READ)
+            if WINDOWS:
+                self.stdout = forward_read_events(self.process.stdout)
+                self.stderr = forward_read_events(self.process.stderr)
+            else:
+                self.stdout = self.process.stdout.fileno()
+                self.stderr = self.process.stderr.fileno()
+            self.loop.add_handler(self.stdout, self.handle_stdout, self.loop.READ)
+            self.loop.add_handler(self.stderr, self.handle_stderr, self.loop.READ)
             self.poller = ioloop.PeriodicCallback(self.poll, self.poll_frequency, self.loop)
             self.poller.start()
             self.notify_start(self.process.pid)
@@ -266,18 +276,29 @@ class LocalProcessLauncher(BaseLauncher):
 
     def signal(self, sig):
         if self.state == 'running':
-            self.process.send_signal(sig)
+            if WINDOWS and sig != SIGINT:
+                # use Windows tree-kill for better child cleanup
+                check_output(['taskkill', '-pid', str(self.process.pid), '-t', '-f'])
+            else:
+                self.process.send_signal(sig)
 
     def interrupt_then_kill(self, delay=2.0):
         """Send INT, wait a delay and then send KILL."""
-        self.signal(SIGINT)
+        try:
+            self.signal(SIGINT)
+        except Exception:
+            self.log.debug("interrupt failed")
+            pass
         self.killer  = ioloop.DelayedCallback(lambda : self.signal(SIGKILL), delay*1000, self.loop)
         self.killer.start()
 
     # callbacks, etc:
     
     def handle_stdout(self, fd, events):
-        line = self.process.stdout.readline()
+        if WINDOWS:
+            line = self.stdout.recv()
+        else:
+            line = self.process.stdout.readline()
         # a stopped process will be readable but return empty strings
         if line:
             self.log.info(line[:-1])
@@ -285,7 +306,10 @@ class LocalProcessLauncher(BaseLauncher):
             self.poll()
     
     def handle_stderr(self, fd, events):
-        line = self.process.stderr.readline()
+        if WINDOWS:
+            line = self.stderr.recv()
+        else:
+            line = self.process.stderr.readline()
         # a stopped process will be readable but return empty strings
         if line:
             self.log.error(line[:-1])
@@ -296,8 +320,8 @@ class LocalProcessLauncher(BaseLauncher):
         status = self.process.poll()
         if status is not None:
             self.poller.stop()
-            self.loop.remove_handler(self.process.stdout.fileno())
-            self.loop.remove_handler(self.process.stderr.fileno())
+            self.loop.remove_handler(self.stdout)
+            self.loop.remove_handler(self.stderr)
             self.notify_stop(dict(exit_code=status, pid=self.process.pid))
         return status
 
@@ -588,10 +612,11 @@ class SSHEngineSetLauncher(LocalEngineSetLauncher):
 
 # This is only used on Windows.
 def find_job_cmd():
-    if os.name=='nt':
+    if WINDOWS:
         try:
             return find_cmd('job')
-        except FindCmdError:
+        except (FindCmdError, ImportError):
+            # ImportError will be raised if win32api is not installed
             return 'job'
     else:
         return 'job'
